web/requirements.txt
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
azure-cognitiveservices-speech==1.24.2
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cookiecutter==2.1.1
cssbeautifier==1.14.6
cssselect==1.1.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.1
django-browser-reload==1.6.0
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-tailwind==3.4.0
djangorestframework==3.14.0
djlint==1.18.0
EditorConfig==0.12.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==4.13.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsbeautifier==1.14.6
lxml==4.9.1
MarkupSafe==2.1.1
media==0.1.6
packaging==21.3
pathspec==0.10.1
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7
PyYAML==6.0
redis==4.3.4
regex==2022.9.13
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.12
wrapt==1.14.1
zipp==3.8.1
